fix uglify
