+++ package.json
@@ -55,7 +55,6 @@
     "subarg": "^1.0.0",
     "temp-path": "^1.0.0",
     "through2": "^2.0.3",
-    "uglify-es": "^3.0.15",
     "uglifyify": "^4.0.0",
     "unassertify": "^2.0.3",
     "watchify": "^3.7.0",
